feat: add required attribute to input
