+++ src/app.d.ts
@@ -1,12 +1,4 @@
 // See https://kit.svelte.dev/docs/types#app
 // for information about these interfaces
 // and what to do when importing types
-declare namespace App {
-	interface Supabase {
-		Database: import('./DatabaseDefinitions').Database;
-		SchemaName: 'public';
-	}
-	interface PageData {
-		session: import('@supabase/supabase-js').Session | null;
-	}
-}
+declare namespace App {}
