+++ src/lib/types/auth.d.ts
@@ -1,0 +1,7 @@
+interface Supabase {
+	Database: import('./DatabaseDefinitions').Database;
+	SchemaName: 'public';
+}
+interface PageData {
+	session: import('@supabase/supabase-js').Session | null;
+}
